Implement like button (#123)
* like button adds on click

* update to comment and like span scss

* isLiked for user implemented. Can like, and update db, and views on index / and resouce/:id pages

+++ routes/api/like.js
@@ -3,25 +3,44 @@ const router = express.Router();
 
 module.exports = (db) => {
   router.post("/", (req, res) => {
-    db.query(
-      `
-      INSERT INTO resource_likes(user_id, resource_id)
-      VALUES(${req.session.userId}, ${req.params.id});
-    `
-    ).then((data) => {
-      res.json({});
-    });
+    console.log(req.session.userId, req.body);
+    if (!req.session.userId || !req.body.resourceId) {
+      res.status(400).json({ error: "Resource id and user id required" });
+    } else {
+      db.query(
+        `
+          INSERT INTO resource_likes(user_id, resource_id)
+          VALUES($1, $2);
+        `,
+        [req.session.userId, req.body.resourceId]
+      )
+        .then((data) => {
+          res.json({});
+        })
+        .catch((err) => {
+          console.log("Error: ", err);
+        });
+    }
   });
 
   router.post("/remove", (req, res) => {
-    db.query(
-      `
-      DELETE FROM resource_likes
-      WHERE user_id = ${req.session.userId} AND resource_id = ${req.params.id};
-    `
-    ).then((data) => {
-      res.json({});
-    });
+    if (!req.session.userId || !req.body.resourceId) {
+      res.status(400).json({ error: "Resource id and user id required" });
+    } else {
+      db.query(
+        `
+        DELETE FROM resource_likes
+        WHERE user_id = $1 AND resource_id = $2;
+        `,
+        [req.session.userId, req.body.resourceId]
+      )
+        .then((data) => {
+          res.json({});
+        })
+        .catch((err) => {
+          console.log("Error: ", err);
+        });
+    }
   });
 
   return router;
